requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
celery==5.5.3
redis==6.4.0
gunicorn==21.2.0
python-dotenv==1.1.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-JWT-Extended==4.7.1
email_validator==2.2.0
PyJWT==2.10.1
SQLAlchemy==2.0.40
WTForms==3.2.1


